# core-common-document-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-document-mgmt
  description: Enterprise Content Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: https://api.getfirefly.io/ecm
    description: Production Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/documents/{id}:
    get:
      tags:
        - Document Controller
      summary: Get document by ID
      description: Returns a document by its ID
      operationId: getDocumentById
      parameters:
        - name: id
          in: path
          description: ID of the document to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved document
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '404':
          description: Document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
    put:
      tags:
        - Document Controller
      summary: Update an existing document
      description: Updates an existing document by its ID
      operationId: updateDocument
      parameters:
        - name: id
          in: path
          description: ID of the document to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentDTO'
        required: true
      responses:
        '200':
          description: Document updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '400':
          description: Invalid document data
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '404':
          description: Document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
    delete:
      tags:
        - Document Controller
      summary: Delete a document
      description: Deletes a document by its ID
      operationId: deleteDocument
      parameters:
        - name: id
          in: path
          description: ID of the document to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Document deleted successfully
        '404':
          description: Document not found
  /api/v1/documents/{documentId}/signatures/{id}:
    get:
      tags:
        - Document Signature Controller
      summary: Get specific signature
      description: Returns a specific signature for a document
      operationId: getDocumentSignature
      parameters:
        - name: documentId
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          description: ID of the signature to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved document signature
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentSignatureDTO'
        '404':
          description: Document signature not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentSignatureDTO'
    put:
      tags:
        - Document Signature Controller
      summary: Update a document signature
      description: Updates an existing signature for a document
      operationId: updateDocumentSignature
      parameters:
        - name: documentId
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          description: ID of the signature to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentSignatureDTO'
        required: true
      responses:
        '200':
          description: Document signature updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentSignatureDTO'
        '400':
          description: Invalid document signature data
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentSignatureDTO'
        '404':
          description: Document signature not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentSignatureDTO'
    delete:
      tags:
        - Document Signature Controller
      summary: Remove a signature from a document
      description: Deletes a signature from a document
      operationId: deleteDocumentSignature
      parameters:
        - name: documentId
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          description: ID of the signature to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Document signature deleted successfully
        '404':
          description: Document signature not found
...
components:
  schemas:
...
    DocumentDTO:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the document
          format: uuid
          readOnly: true
        name:
          type: string
          description: Name of the document
        description:
          type: string
          description: Description of the document
        fileName:
          type: string
          description: Original file name of the document
        fileExtension:
          type: string
          description: File extension of the document
        mimeType:
          type: string
          description: MIME type of the document
        fileSize:
          type: integer
          description: Size of the document file in bytes
          format: int64
        documentType:
          type: string
          description: Type of the document
          enum:
            - DOCUMENT
            - IMAGE
            - PDF
            - CONTRACT
            - INVOICE
            - RECEIPT
            - FORM
            - REPORT
            - EMAIL
            - AUDIO
            - VIDEO
            - ARCHIVE
            - OTHER
        documentStatus:
          type: string
          description: Current status of the document
          enum:
            - DRAFT
            - UNDER_REVIEW
            - APPROVED
            - REJECTED
            - PUBLISHED
            - ARCHIVED
            - MARKED_FOR_DELETION
            - DELETED
            - LOCKED
            - EXPIRED
        storageType:
          type: string
          description: Storage type of the document
          enum:
            - LOCAL_FILESYSTEM
            - DATABASE
            - S3
            - AZURE_BLOB
            - GOOGLE_CLOUD_STORAGE
            - CDN
            - DISTRIBUTED_FS
            - EXTERNAL_REFERENCE
        storagePath:
          type: string
          description: Path where the document is stored
        securityLevel:
          type: string
          description: Security level of the document
          enum:
            - PUBLIC
            - INTERNAL
            - CONFIDENTIAL
            - RESTRICTED
            - SECRET
            - TOP_SECRET
        folderId:
          type: string
          description: ID of the folder containing the document
          format: uuid
        isEncrypted:
          type: boolean
          description: Indicates if the document is encrypted
        isIndexed:
          type: boolean
          description: Indicates if the document is indexed for search
        isLocked:
          type: boolean
          description: Indicates if the document is locked for editing
        lockedBy:
          type: string
          description: User who locked the document
        lockedUntil:
          type: string
          description: Date and time until the document is locked
          format: date-time
        expirationDate:
          type: string
          description: Date and time when the document expires
          format: date-time
        retentionDate:
          type: string
          description: Date and time until the document should be retained
          format: date-time
        tenantId:
          type: string
          description: Tenant ID for multi-tenancy support
        createdAt:
          type: string
          description: Date and time when the document was created
          format: date-time
        createdBy:
          type: string
          description: User who created the document
        updatedAt:
          type: string
          description: Date and time when the document was last updated
          format: date-time
        updatedBy:
          type: string
          description: User who last updated the document
        version:
          type: integer
          description: Version number for optimistic locking
          format: int64
        checksum:
          type: string
          description: Checksum of the document file for integrity verification
      description: Document data transfer object
    DocumentSignatureDTO:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the document signature
          format: uuid
          readOnly: true
        documentId:
          type: string
          description: ID of the document being signed
          format: uuid
        documentVersionId:
          type: string
          description: ID of the document version being signed
          format: uuid
        signatureProviderId:
          type: string
          description: ID of the signature provider
          format: uuid
        signerPartyId:
          type: string
          description: ID of the party signing the document
          format: uuid
        signerName:
          type: string
          description: Name of the signer
        signerEmail:
          type: string
          description: Email of the signer
        signatureType:
          type: string
          description: Type of signature
          enum:
            - BASIC
            - ADVANCED
            - ADVANCED_WITH_QCERT
            - QUALIFIED
            - DIGITAL
            - BIOMETRIC
        signatureFormat:
          type: string
          description: Format of signature
          enum:
            - PADES
            - XADES
            - CADES
            - JADES
            - PDF_VISIBLE
            - PDF_INVISIBLE
            - TIMESTAMP
        signatureStatus:
          type: string
          description: Status of the signature
          enum:
            - PENDING
            - IN_PROGRESS
            - SIGNED
            - REJECTED
            - EXPIRED
            - REVOKED
            - FAILED
            - CANCELED
        signatureData:
          type: string
          description: Signature data (base64 encoded)
        signatureCertificate:
          type: string
          description: Signature certificate (base64 encoded)
        signaturePositionX:
          type: integer
          description: X position of the signature on the document
          format: int32
        signaturePositionY:
          type: integer
          description: Y position of the signature on the document
          format: int32
        signaturePage:
          type: integer
          description: Page number where the signature appears
          format: int32
        signatureWidth:
          type: integer
          description: Width of the signature
          format: int32
        signatureHeight:
          type: integer
          description: Height of the signature
          format: int32
        signatureReason:
          type: string
          description: Reason for signing
        signatureLocation:
          type: string
          description: Location where the document was signed
        signatureContactInfo:
          type: string
          description: Contact information of the signer
        expirationDate:
          type: string
          description: Date and time when the signature expires
          format: date-time
        signedAt:
          type: string
          description: Date and time when the document was signed
          format: date-time
        tenantId:
          type: string
          description: Tenant ID for multi-tenancy support
        customSignatureMessage:
          maxLength: 1000
          type: string
          description: Custom message displayed to the signer during the signing process. This message will be shown in the ECM provider's interface. If not provided, defaults to configuration value.
          example: Please review and sign this important contract by Friday
        signerLanguage:
          pattern: ^[a-z]{2}$
          type: string
          description: 'ISO 639-1 language code for the signer''s preferred language. Determines the language of the ECM provider''s signing interface. Supported values: en, es, fr, de, it, pt, etc.'
          example: en
          default: en
        signerTimeZone:
          type: string
          description: Time zone identifier for the signer. Used for displaying dates and times in the signer's local time zone during the signing process.
          example: America/New_York
          default: UTC
        signingOrder:
          maximum: 100
          minimum: 1
          type: integer
          description: Signing order for multiple signers. Determines the sequence in which signers must complete their signatures. Lower numbers sign first.
          format: int32
          example: 1
          default: 1
        signerRole:
          type: string
          description: 'Role of the signer in the document workflow. Affects permissions and display in the ECM provider interface. Valid values: Signer, Approver, Reviewer, Witness, Notary, CC'
          example: Approver
          default: Signer
          enum:
            - Signer
            - Approver
            - Reviewer
            - Witness
            - Notary
            - CC
        signatureRequired:
          type: boolean
          description: Whether this signature is required for document completion. If false, the signature is optional and document can be completed without it.
          example: true
          default: true
        authenticationMethod:
          type: string
          description: 'Authentication method required before signing. Determines how the signer must verify their identity before accessing the document. Valid values: EMAIL, SMS, PHONE, ACCESS_CODE, ID_CHECK, NONE'
          example: EMAIL
          default: EMAIL
          enum:
            - EMAIL
            - SMS
            - PHONE
            - ACCESS_CODE
            - ID_CHECK
            - NONE
        externalSignerId:
          type: string
          description: External signer ID assigned by the ECM provider (e.g., DocuSign envelope ID). This field is populated automatically when the signature request is created through an ECM provider and is used for tracking and status updates.
          readOnly: true
          example: envelope-12345-abcdef
        signingUrl:
          type: string
          description: Direct URL for the signer to access the document for signing. Generated by the ECM provider and sent to the signer via email or other means. This URL is typically time-limited and signer-specific.
          readOnly: true
          example: https://demo.docusign.net/signing/startinsession.aspx?t=12345
        ecmMetadata:
          maxLength: 5000
          type: string
          description: Additional metadata specific to the ECM provider. Can contain provider-specific configuration, tracking information, or custom fields in JSON format. Maximum 5000 characters.
          example: '{"provider":"docusign","templateId":"template-123","customFields":{"department":"legal"}}'
        createdAt:
          type: string
          description: Date and time when the document signature was created
          format: date-time
          readOnly: true
        createdBy:
          type: string
          description: User who created the document signature
          readOnly: true
        updatedAt:
          type: string
          description: Date and time when the document signature was last updated
          format: date-time
          readOnly: true
        updatedBy:
          type: string
          description: User who last updated the document signature
          readOnly: true
        version:
          type: integer
          description: Version number for optimistic locking
          format: int64
          readOnly: true
      description: Document signature data transfer object
...